Set "virtualmac: no" for carp to use the MAC address of the parent.
The cloud doesn't allow multiple MAC addresses (as the CARP address)
on the same virtual switch.  Esdenera's "virtualmac: no" option allows
to inherit the parent's MAC address instead.  This means that a
failover will change the firewall MAC address, but most clouds seems
to handle that fine (with a little extra delay).